Silenced a compiler warning in ntfswipe (cosmetic)
Avoid a signed-to-unsigned compare warning

+++ ntfsprogs/ntfswipe.c
@@ -981,7 +981,7 @@ static s64 wipe_mft(ntfs_volume *vol, int byte, enum action act)
 			// We know that the end marker will only take 4 bytes
 			size = le32_to_cpu(rec->bytes_in_use) - 4;
 
-			if ((size <= 0) || (size > vol->mft_record_size)) {
+			if ((size <= 0) || (size > (int)vol->mft_record_size)) {
 				ntfs_log_error("Bad mft record %lld\n",
 						(long long)i);
 				total = -1;
